backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
sentence-transformers==2.3.1
numpy==1.26.3
python-multipart==0.0.6
aiofiles==23.2.1
chromadb==0.4.22
